# backend/storage/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Civic Intelligence Platform API'
  description: |
    REST API for the Civic Intelligence Platform.
    - M2 adds the Authentication namespace (citizen OTP login, JWT issuance, refresh rotation).
    - M3 adds the Super Admin master-data CRUD namespaces (departments, settings, feature flags) and the geography tree that every portal reads on cold start.
  contact:
    name: 'Civic Platform Engineering'
    email: eng@cip.local
  license:
    name: Proprietary
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8080'
    description: 'Local docker-compose stack'
  -
    url: '{scheme}://{host}'
    description: 'Configurable host'
    variables:
      scheme:
        enum:
          - http
          - https
        default: https
      host:
        default: api.cip.local
paths:
...
  /api/v1/admin/departments/{department}/officers/{user}:
    delete:
      tags: ['Department Admin']
      summary: 'Detach an officer from a department (M11)'
      operationId: admin.departments.officers.destroy
      parameters:
        - $ref: '#/components/parameters/PathDepartmentId'
        - in: path
          name: user
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DetachOfficerResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: 'Rotate the refresh token and issue a new access token'
      description: |
        Consumes a one-time refresh token and returns a new access token
        plus a new refresh token. The presented refresh token is marked
        used and cannot be presented again. Replay attempts emit a
        security event (REFRESH_TOKEN_REPLAY) and revoke the entire
        session chain.
      operationId: auth.refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 'Tokens rotated successfully.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/admin/routing-rules/{rule}:
    parameters:
      - in: path
        name: rule
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Routing]
      summary: 'Show a routing rule'
      operationId: admin.routing-rules.show
      responses:
        '200':
          description: 'Routing rule found.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoutingRuleResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Routing]
      summary: 'Update a routing rule'
      operationId: admin.routing-rules.update
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoutingRuleUpdateRequest' }
      responses:
        '200':
          description: 'Routing rule updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoutingRuleResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Routing]
      summary: 'Soft-delete a routing rule'
      operationId: admin.routing-rules.destroy
      responses:
        '200':
          description: 'Routing rule soft-deleted.'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted: { type: boolean, example: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/users/{user}:
    get:
      tags: ['Super Admin']
      summary: 'Show a user (super_admin only)'
      operationId: admin.users.show
      parameters:
        - in: path
          name: user
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: ['Super Admin']
      summary: 'Update a user (super_admin only)'
      operationId: admin.users.update
      parameters:
        - in: path
          name: user
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: 'User updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: ['Super Admin']
      summary: 'Soft-delete a user (super_admin only)'
      operationId: admin.users.destroy
      parameters:
        - in: path
          name: user
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'User deleted.'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/roles/{role}:
    get:
      tags: ['Super Admin']
      summary: 'Show a role (super_admin only)'
      operationId: admin.roles.show
      parameters:
        - in: path
          name: role
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 'Role details.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: ['Super Admin']
      summary: 'Update a role (super_admin only)'
      operationId: admin.roles.update
      parameters:
        - in: path
          name: role
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleUpdateRequest' }
      responses:
        '200':
          description: 'Role updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RoleResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: ['Super Admin']
      summary: 'Delete a role (super_admin only; protected roles forbidden)'
      operationId: admin.roles.destroy
      parameters:
        - in: path
          name: role
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 'Role deleted.'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/permissions/{permission}:
    get:
      tags: ['Super Admin']
      summary: 'Show a permission (super_admin only)'
      operationId: admin.permissions.show
      parameters:
        - in: path
          name: permission
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 'Permission details.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PermissionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: ['Super Admin']
      summary: 'Delete a permission (super_admin only)'
      operationId: admin.permissions.destroy
      parameters:
        - in: path
          name: permission
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 'Permission deleted.'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/report-types/{report_type}:
    get:
      tags: ['Super Admin']
      summary: 'Show a report type (super_admin only)'
      operationId: admin.report-types.show
      parameters:
        - in: path
          name: report_type
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'Report type details.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportTypeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: ['Super Admin']
      summary: 'Update a report type (super_admin only)'
      operationId: admin.report-types.update
      parameters:
        - in: path
          name: report_type
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportTypeUpdateRequest' }
      responses:
        '200':
          description: 'Report type updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportTypeResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: ['Super Admin']
      summary: 'Soft-delete a report type (super_admin only)'
      operationId: admin.report-types.destroy
      parameters:
        - in: path
          name: report_type
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'Report type deleted.'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/security-policies/{key}:
    get:
      tags: ['Super Admin']
      summary: 'Show a security policy by key (super_admin only)'
      operationId: admin.security-policies.show
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 'Policy details.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SecurityPolicyResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: ['Super Admin']
      summary: 'Update a security policy by key (super_admin only)'
      operationId: admin.security-policies.update
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecurityPolicyUpsertRequest' }
      responses:
        '200':
          description: 'Policy updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SecurityPolicyResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: ['Super Admin']
      summary: 'Delete a security policy by key (super_admin only)'
      operationId: admin.security-policies.destroy
      parameters:
        - in: path
          name: key
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 'Policy deleted.'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/admin/integrations/{integration}:
    get:
      tags: [Integrations]
      summary: 'Show one integration (super_admin only)'
      operationId: admin.integrations.show
      parameters:
        - in: path
          name: integration
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'Integration row (credentials masked).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Integrations]
      summary: 'Update an integration (super_admin only)'
      operationId: admin.integrations.update
      parameters:
        - in: path
          name: integration
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IntegrationUpdateRequest' }
      responses:
        '200':
          description: 'Integration updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Integrations]
      summary: 'Soft-delete an integration (super_admin only)'
      operationId: admin.integrations.destroy
      parameters:
        - in: path
          name: integration
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 'Integration deleted.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/admin/notification-configs/{config}:
    get:
      tags: [NotificationConfigs]
      summary: 'Show one channel config (super_admin only)'
      operationId: admin.notification-configs.show
      parameters:
        - in: path
          name: config
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'Channel config (credentials masked).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationChannelConfigResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [NotificationConfigs]
      summary: 'Update a channel config (super_admin only)'
      operationId: admin.notification-configs.update
      parameters:
        - in: path
          name: config
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NotificationChannelConfigUpdateRequest' }
      responses:
        '200':
          description: 'Channel config updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationChannelConfigResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [NotificationConfigs]
      summary: 'Soft-delete a channel config (super_admin only)'
      operationId: admin.notification-configs.destroy
      parameters:
        - in: path
          name: config
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 'Channel config deleted.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/admin/organizations/{organization}:
    get:
      tags: [Organizations]
      summary: 'Show one organization (super_admin only)'
      operationId: admin.organizations.show
      parameters:
        - in: path
          name: organization
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 'Organization row.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Organizations]
      summary: 'Update an organization (super_admin only)'
      operationId: admin.organizations.update
      parameters:
        - in: path
          name: organization
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationUpdateRequest' }
      responses:
        '200':
          description: 'Organization updated.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Organizations]
      summary: 'Soft-delete an organization (super_admin only)'
      operationId: admin.organizations.destroy
      parameters:
        - in: path
          name: organization
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: 'Organization deleted.' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    sanctum:
      type: http
      description: 'Laravel Sanctum personal access token (M2).'
      bearerFormat: 'Sanctum PAT'
      scheme: bearer
  parameters:
...
    PathDepartmentId:
      in: path
      name: department
      required: true
      schema: { type: string, format: uuid }
  schemas:
    ApiResponse:
      type: object
      description: 'Standard API envelope (docs/03 §20).'
      required: [success, message, data, code, trace_id]
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { type: object, additionalProperties: true, nullable: true }
        errors: { type: object, additionalProperties: true, nullable: true }
        code: { type: string, description: 'Application-specific error code; null on success.' }
        trace_id: { type: string, description: 'Request id used for cross-service correlation.' }
        meta: { type: object, additionalProperties: true, nullable: true }
...
    ErrorResponse:
      type: object
      required: [success, message, code, trace_id]
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 'Invalid OTP code.' }
        code: { type: string, example: 'OTP_INVALID' }
        trace_id: { type: string, example: '0HMM8M3N5G2F1:00000001' }
        errors: { type: object, additionalProperties: true, nullable: true }
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string, minLength: 32, description: 'Opaque refresh token returned by /verify-otp or a previous /refresh.' }
    RefreshTokenResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required: [token, refresh_token, refresh_expires_at]
              properties:
                token:
                  type: object
                  required: [access_token, type, expires_at]
                  properties:
                    access_token: { type: string }
                    type: { type: string, example: 'Bearer' }
                    expires_at: { type: string, format: date-time, nullable: true }
                refresh_token: { type: string }
                refresh_expires_at: { type: string, format: date-time }
...
    User:
      type: object
      description: 'UserResource payload (docs/05 §3). Roles and permissions are populated when the relation was eager-loaded.'
      properties:
        id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        mobile: { type: string, example: '9876543210' }
        email: { type: string, format: email, nullable: true }
        anonymous_enabled: { type: boolean }
        status: { type: string, enum: [active, suspended, disabled, pending] }
        otp_verified_at: { type: string, format: date-time, nullable: true }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        roles:
          type: array
          items: { type: string, example: 'citizen' }
        permissions:
          type: array
          items: { type: string, example: 'reports.submit' }
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/User' }
...
    UserUpdateRequest:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        mobile: { type: string, maxLength: 32 }
        email: { type: string, format: email, nullable: true, maxLength: 255 }
        password: { type: string, nullable: true, minLength: 8, maxLength: 128 }
        status:
          type: string
          nullable: true
          enum: ['active', 'suspended', 'banned', 'pending']
        anonymous_enabled: { type: boolean, nullable: true }
        roles:
          type: array
          nullable: true
          items: { type: string }
    RoleResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Role' }
...
    Role:
      type: object
      properties:
        id: { type: integer, example: 1 }
        name: { type: string, example: 'moderator' }
        guard_name: { type: string, example: 'web' }
        protected: { type: boolean, example: false }
        permissions:
          type: array
          items: { type: string, example: 'reports.view' }
        created_at: { type: string, format: date-time, nullable: true }
...
    RoleUpdateRequest:
      type: object
      properties:
        name: { type: string, maxLength: 64 }
        guard_name: { type: string, nullable: true, maxLength: 64 }
        permissions:
          type: array
          nullable: true
          items: { type: string }
...
    PermissionResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Permission' }
...
    Permission:
      type: object
      properties:
        id: { type: integer, example: 12 }
        name: { type: string, example: 'reports.view' }
        guard_name: { type: string, example: 'web' }
        created_at: { type: string, format: date-time, nullable: true }
    SecurityPolicyResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/SecurityPolicy' }
...
    SecurityPolicy:
      type: object
      properties:
        id: { type: string, format: uuid }
        key: { type: string, example: 'password.min_length' }
        value: { type: object, nullable: true, additionalProperties: true }
        type: { type: string, enum: ['string', 'int', 'bool', 'array'] }
        description: { type: string, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
    SecurityPolicyUpsertRequest:
      type: object
      required: [key]
      properties:
        key: { type: string, maxLength: 128, pattern: '^[a-z0-9._]+$' }
        value: { type: object, nullable: true, additionalProperties: true }
        type: { type: string, nullable: true, enum: ['string', 'int', 'bool', 'array'] }
        description: { type: string, nullable: true, maxLength: 255 }
...
    IntegrationResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Integration' }
...
    IntegrationUpdateRequest:
      type: object
      properties:
        code: { type: string, pattern: '^[a-z0-9_]+$', maxLength: 64 }
        provider: { type: string, maxLength: 64 }
        display_name: { type: string, maxLength: 128 }
        base_url: { type: string, format: uri, maxLength: 512 }
        credentials: { type: object, additionalProperties: true }
        settings: { type: object, nullable: true }
        status: { type: string, enum: [active, degraded, disabled] }
...
    NotificationChannelConfigResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/NotificationChannelConfig' }
...
    NotificationChannelConfigUpdateRequest:
      type: object
      properties:
        display_name: { type: string, maxLength: 128 }
        credentials: { type: object, additionalProperties: true }
        retry_policy:
          type: object
          nullable: true
          properties:
            tries: { type: integer, minimum: 1, maximum: 10 }
            backoff:
              type: array
              items: { type: integer, minimum: 1, maximum: 86400 }
              maxItems: 10
        settings: { type: object, nullable: true }
        per_locale_defaults: { type: object, nullable: true }
        active: { type: boolean }
...
    OrganizationResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/Organization' }
...
    OrganizationUpdateRequest:
      type: object
      properties:
        code: { type: string, pattern: '^[a-z0-9_-]+$', maxLength: 64 }
        name: { type: string, maxLength: 128 }
        legal_name: { type: string, nullable: true, maxLength: 255 }
        domain: { type: string, nullable: true, maxLength: 128 }
        contact: { type: object, nullable: true }
        branding: { type: object, nullable: true }
        storage_quota_mb: { type: integer, minimum: 0, maximum: 1048576 }
        settings: { type: object, nullable: true }
        active: { type: boolean }
...
    ReportTypeResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ReportType' }
...
    ReportTypeUpdateRequest:
      type: object
      properties:
        name: { type: string, maxLength: 255 }
        code: { type: string, maxLength: 64, pattern: '^[a-z0-9_]+$' }
        description: { type: string, nullable: true, maxLength: 1000 }
        icon: { type: string, nullable: true, maxLength: 64 }
        color: { type: string, nullable: true, maxLength: 16 }
        department_default_id: { type: string, format: uuid, nullable: true }
        requires_video: { type: boolean }
        requires_photo: { type: boolean }
        min_photos: { type: integer, minimum: 0, maximum: 99 }
        max_photos: { type: integer, minimum: 0, maximum: 99 }
        workflow_definition_id: { type: string, format: uuid, nullable: true }
        validation_rules: { type: object, nullable: true, additionalProperties: true }
        active: { type: boolean }
...
  responses:
    NotFound:
      description: 'Resource not found.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 'Authenticated but the role is not permitted.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: 'Request body failed validation.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 'Authentication required or invalid.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...